Moves DashConfig to own file. Adds XML comments.
